+++ sawp-diameter/src/lib.rs
@@ -5,6 +5,7 @@ use sawp::error::{NomError, Result};
 use sawp::parser::{Direction, Parse};
 use sawp::probe::Probe;
 use sawp::protocol::Protocol;
+use sawp_flags::{BitFlags, Flag, Flags};
 
 use nom::bytes::streaming::tag;
 use nom::bytes::streaming::take;
@@ -17,8 +18,6 @@ use nom::IResult;
 use num_enum::TryFromPrimitive;
 use std::convert::TryFrom;
 use std::net::{IpAddr, Ipv4Addr, Ipv6Addr};
-
-use bitflags::bitflags;
 
 #[derive(Debug)]
 pub struct Diameter {}
@@ -128,10 +127,13 @@ pub enum Value {
 }
 
 impl Value {
-    pub fn new<'a>(code: &AttributeCode, data: &'a [u8]) -> IResult<&'a [u8], (Self, ErrorFlags)> {
+    pub fn new<'a>(
+        code: &AttributeCode,
+        data: &'a [u8],
+    ) -> IResult<&'a [u8], (Self, Flags<ErrorFlags>)> {
         match code {
             AttributeCode::AcctSessionId | AttributeCode::ProxyState => {
-                Ok((&[], (Value::OctetString(data.into()), ErrorFlags::NONE)))
+                Ok((&[], (Value::OctetString(data.into()), ErrorFlags::none())))
             }
             AttributeCode::AcctInterimInterval
             | AttributeCode::AccountingRecordNumber
@@ -151,11 +153,11 @@ impl Value {
             | AttributeCode::SupportedVendorId
             | AttributeCode::VendorId => {
                 let (input, val) = be_u32(data)?;
-                Ok((input, (Value::Unsigned32(val), ErrorFlags::NONE)))
+                Ok((input, (Value::Unsigned32(val), ErrorFlags::none())))
             }
             AttributeCode::AccountingSubSessionId => {
                 let (input, val) = be_u64(data)?;
-                Ok((input, (Value::Unsigned64(val), ErrorFlags::NONE)))
+                Ok((input, (Value::Unsigned64(val), ErrorFlags::none())))
             }
             AttributeCode::AccountingRealtimeRequired
             | AttributeCode::AccountingRecordType
@@ -167,7 +169,7 @@ impl Value {
             | AttributeCode::SessionServerFailover
             | AttributeCode::TerminationCause => {
                 let (input, val) = be_u32(data)?;
-                Ok((input, (Value::Enumerated(val), ErrorFlags::NONE)))
+                Ok((input, (Value::Enumerated(val), ErrorFlags::none())))
             }
             AttributeCode::ExperimentalResult
             | AttributeCode::FailedAVP
@@ -182,8 +184,11 @@ impl Value {
             | AttributeCode::ProductName
             | AttributeCode::SessionId
             | AttributeCode::UserName => match String::from_utf8(data.to_vec()) {
-                Ok(string) => Ok((&[], (Value::UTF8String(string), ErrorFlags::NONE))),
-                Err(_) => Ok((&[], (Value::Unhandled(data.into()), ErrorFlags::DATA_VALUE))),
+                Ok(string) => Ok((&[], (Value::UTF8String(string), ErrorFlags::none()))),
+                Err(_) => Ok((
+                    &[],
+                    (Value::Unhandled(data.into()), ErrorFlags::DataValue.into()),
+                )),
             },
             AttributeCode::DestinationHost
             | AttributeCode::DestinationRealm
@@ -192,12 +197,18 @@ impl Value {
             | AttributeCode::OriginRealm
             | AttributeCode::ProxyHost
             | AttributeCode::RouteRecord => match String::from_utf8(data.to_vec()) {
-                Ok(string) => Ok((&[], (Value::DiameterIdentity(string), ErrorFlags::NONE))),
-                Err(_) => Ok((&[], (Value::Unhandled(data.into()), ErrorFlags::DATA_VALUE))),
+                Ok(string) => Ok((&[], (Value::DiameterIdentity(string), ErrorFlags::none()))),
+                Err(_) => Ok((
+                    &[],
+                    (Value::Unhandled(data.into()), ErrorFlags::DataValue.into()),
+                )),
             },
             AttributeCode::RedirectHost => match String::from_utf8(data.to_vec()) {
-                Ok(string) => Ok((&[], (Value::DiameterURI(string), ErrorFlags::NONE))),
-                Err(_) => Ok((&[], (Value::Unhandled(data.into()), ErrorFlags::DATA_VALUE))),
+                Ok(string) => Ok((&[], (Value::DiameterURI(string), ErrorFlags::none()))),
+                Err(_) => Ok((
+                    &[],
+                    (Value::Unhandled(data.into()), ErrorFlags::DataValue.into()),
+                )),
             },
             AttributeCode::HostIPAddress => match data.len() {
                 4 => Ok((
@@ -207,7 +218,7 @@ impl Value {
                         Value::Address(IpAddr::V4(Ipv4Addr::from(
                             <[u8; 4]>::try_from(data).unwrap(),
                         ))),
-                        ErrorFlags::NONE,
+                        ErrorFlags::none(),
                     ),
                 )),
                 16 => Ok((
@@ -217,19 +228,19 @@ impl Value {
                         Value::Address(IpAddr::V6(Ipv6Addr::from(
                             <[u8; 16]>::try_from(data).unwrap(),
                         ))),
-                        ErrorFlags::NONE,
+                        ErrorFlags::none(),
                     ),
                 )),
                 _ => Ok((
                     &[],
-                    (Value::Unhandled(data.into()), ErrorFlags::DATA_LENGTH),
+                    (Value::Unhandled(data.into()), ErrorFlags::DataLength.into()),
                 )),
             },
             AttributeCode::EventTimestamp => {
                 let (input, seconds) = be_u32(data)?;
-                Ok((input, (Value::Time(seconds), ErrorFlags::NONE)))
+                Ok((input, (Value::Time(seconds), ErrorFlags::none())))
             }
-            _ => Ok((&[], (Value::Unhandled(data.into()), ErrorFlags::NONE))),
+            _ => Ok((&[], (Value::Unhandled(data.into()), ErrorFlags::none()))),
         }
     }
 }
@@ -244,25 +255,24 @@ pub struct AVP {
     padding: Vec<u8>,
 }
 
-bitflags! {
-    /// Flags identify messages which parse successfully
-    /// but contain invalid data. The caller can use the message's
-    /// error flags to see if and what errors were in the
-    /// pack of bytes and take action using this information.
-    pub struct ErrorFlags: u8 {
-        const NONE = 0b0000_0000;
-        const DATA_VALUE = 0b0000_0001;
-        const DATA_LENGTH = 0b0000_0010;
-        const NON_ZERO_RESERVED = 0b0000_0100;
-        const NON_ZERO_PADDING = 0b0000_1000;
-    }
+/// Flags identify messages which parse successfully
+/// but contain invalid data. The caller can use the message's
+/// error flags to see if and what errors were in the
+/// pack of bytes and take action using this information.
+#[repr(u8)]
+#[derive(Clone, Copy, Debug, PartialEq, Eq, BitFlags)]
+pub enum ErrorFlags {
+    DataValue = 0b0000_0001,
+    DataLength = 0b0000_0010,
+    NonZeroReserved = 0b0000_0100,
+    NonZeroPadding = 0b0000_1000,
 }
 
 #[derive(Debug, PartialEq)]
 pub struct Message {
     pub header: Header,
     pub avps: Vec<AVP>,
-    pub error_flags: ErrorFlags,
+    pub error_flags: Flags<ErrorFlags>,
 }
 
 /// Create a parser to read diameter length and ensure input is long enough
@@ -344,8 +354,8 @@ impl Header {
         (self.length as usize) - Self::SIZE
     }
 
-    pub fn parse(input: &[u8]) -> IResult<&[u8], (Self, ErrorFlags)> {
-        let mut error_flags = ErrorFlags::NONE;
+    pub fn parse(input: &[u8]) -> IResult<&[u8], (Self, Flags<ErrorFlags>)> {
+        let mut error_flags = ErrorFlags::none();
         let (input, version) = tag(&[1u8])(input)?;
         let (input, length) = length(Self::PRE_LENGTH_SIZE)(input)?;
         if (length as usize) < Self::SIZE {
@@ -356,7 +366,7 @@ impl Header {
         }
         let (input, flags) = be_u8(input)?;
         if Self::reserved_set(flags) {
-            error_flags |= ErrorFlags::NON_ZERO_RESERVED;
+            error_flags |= ErrorFlags::NonZeroReserved;
         }
         let (input, code) = be_u24(input)?;
         let (input, app_id) = be_u32(input)?;
@@ -434,12 +444,12 @@ impl AVP {
         self.flags & Self::RESERVED_MASK
     }
 
-    pub fn parse(input: &[u8]) -> IResult<&[u8], (Self, ErrorFlags)> {
-        let mut error_flags = ErrorFlags::NONE;
+    pub fn parse(input: &[u8]) -> IResult<&[u8], (Self, Flags<ErrorFlags>)> {
+        let mut error_flags = ErrorFlags::none();
         let (input, raw_code) = be_u32(input)?;
         let (input, flags) = be_u8(input)?;
         if Self::reserved_set(flags) {
-            error_flags |= ErrorFlags::NON_ZERO_RESERVED;
+            error_flags |= ErrorFlags::NonZeroReserved;
         }
         let (input, length) = length(Self::PRE_LENGTH_SIZE)(input)?;
         let header_size = if Self::vendor_specific_flag(flags) {
@@ -464,13 +474,13 @@ impl AVP {
         let (input, data) = take(data_length)(input)?;
         let (input, padding) = take(Self::padding(data_length))(input)?;
         if !padding.iter().all(|&item| item == 0) {
-            error_flags |= ErrorFlags::NON_ZERO_PADDING;
+            error_flags |= ErrorFlags::NonZeroPadding;
         }
         let attribute = Attribute::new(raw_code);
         let value = match Value::new(&attribute.code, data) {
             Ok((rest, (value, flags))) => {
                 if !rest.is_empty() {
-                    error_flags |= ErrorFlags::DATA_LENGTH;
+                    error_flags |= ErrorFlags::DataLength;
                 }
                 error_flags |= flags;
                 value
@@ -480,11 +490,11 @@ impl AVP {
                 code: ErrorKind::LengthValue,
             }))
             | Err(nom::Err::Incomplete(_)) => {
-                error_flags |= ErrorFlags::DATA_LENGTH;
+                error_flags |= ErrorFlags::DataLength;
                 Value::Unhandled(data.into())
             }
             Err(_) => {
-                error_flags |= ErrorFlags::DATA_VALUE;
+                error_flags |= ErrorFlags::DataValue;
                 Value::Unhandled(data.into())
             }
         };
@@ -506,12 +516,6 @@ impl AVP {
     }
 }
 
-impl std::fmt::Display for ErrorFlags {
-    fn fmt(&self, fmt: &mut std::fmt::Formatter) -> std::fmt::Result {
-        write!(fmt, "{:?}", self)
-    }
-}
-
 impl Protocol<'_> for Diameter {
     type Message = Message;
 
@@ -520,13 +524,13 @@ impl Protocol<'_> for Diameter {
     }
 }
 
-fn parse_avps(input: &[u8]) -> IResult<&[u8], (Vec<AVP>, ErrorFlags)> {
+fn parse_avps(input: &[u8]) -> IResult<&[u8], (Vec<AVP>, Flags<ErrorFlags>)> {
     let (rest, avps_flags) = many0(combinator::complete(AVP::parse))(input)?;
     if !rest.is_empty() {
         // many0 will stop if subparser fails, but should read all
         Err(nom::Err::Error(NomError::new(input, ErrorKind::Many0)))
     } else {
-        let mut error_flags = ErrorFlags::NONE;
+        let mut error_flags = ErrorFlags::none();
         let mut avps = Vec::new();
         for (avp, flag) in avps_flags {
             error_flags |= flag;
@@ -543,7 +547,7 @@ impl<'a> Parse<'a> for Diameter {
         input: &'a [u8],
         _direction: Direction,
     ) -> Result<(&'a [u8], Option<Self::Message>)> {
-        let mut error_flags = ErrorFlags::NONE;
+        let mut error_flags = ErrorFlags::none();
         let (input, (header, flags)) = Header::parse(input)?;
         error_flags |= flags;
 
@@ -643,7 +647,7 @@ mod tests {
                     hop_id: 0x53ca_fe6a,
                     end_id: 0x7dc0_a11b,
                 },
-                ErrorFlags::NONE,
+                ErrorFlags::none(),
             )))
         ),
         case::reserved_set(
@@ -674,7 +678,7 @@ mod tests {
                     hop_id: 0x53ca_fe6a,
                     end_id: 0x7dc0_a11b,
                 },
-                ErrorFlags::NON_ZERO_RESERVED,
+                ErrorFlags::NonZeroReserved.into(),
             )))
         ),
         case::diagnostic(
@@ -697,7 +701,7 @@ mod tests {
             Err(nom::Err::Incomplete(nom::Needed::new(4)))
         ),
     )]
-    fn test_header(input: &[u8], expected: IResult<&[u8], (Header, ErrorFlags)>) {
+    fn test_header(input: &[u8], expected: IResult<&[u8], (Header, Flags<ErrorFlags>)>) {
         assert_eq!(Header::parse(input), expected);
     }
 
@@ -733,7 +737,7 @@ mod tests {
                     value: Value::DiameterIdentity("backend.eap.testbed.aaa".into()),
                     padding: vec![0x00],
                 },
-                ErrorFlags::NONE,
+                ErrorFlags::none(),
             )))
         ),
         case::diagnostic_vendor_id(
@@ -762,7 +766,7 @@ mod tests {
                     value: Value::DiameterIdentity("".into()),
                     padding: Vec::new(),
                 },
-                ErrorFlags::NONE,
+                ErrorFlags::none(),
             )))
         ),
         case::unsigned_32_format(
@@ -794,7 +798,7 @@ mod tests {
 
                     padding: vec![0x00, 0x00, 0x00],
                 },
-                ErrorFlags::DATA_LENGTH,
+                ErrorFlags::DataLength.into(),
             )))
         ),
         case::unsigned_64_format(
@@ -823,7 +827,7 @@ mod tests {
                     value: Value::Unsigned64(528_297_886_211),
                     padding: Vec::new(),
                 },
-                ErrorFlags::NONE,
+                ErrorFlags::none(),
             )))
         ),
         case::enumerated_format(
@@ -851,7 +855,7 @@ mod tests {
                     value: Value::Enumerated(2),
                     padding: Vec::new(),
                 },
-                ErrorFlags::NONE,
+                ErrorFlags::none(),
             )))
         ),
         case::octet_string_format(
@@ -882,7 +886,7 @@ mod tests {
                     value: Value::OctetString(vec![0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07]),
                     padding: vec![0xef],
                 },
-                ErrorFlags::NON_ZERO_PADDING,
+                ErrorFlags::NonZeroPadding.into(),
             )))
         ),
         case::utf8_string_format(
@@ -912,7 +916,7 @@ mod tests {
                     value: Value::UTF8String("Hello World!".into()),
                     padding: Vec::new(),
                 },
-                ErrorFlags::NONE,
+                ErrorFlags::none(),
             )))
         ),
         case::diameter_uri_format(
@@ -944,7 +948,7 @@ mod tests {
                     value: Value::DiameterURI("example.com".into()),
                     padding: vec![0x00],
                 },
-                ErrorFlags::NONE,
+                ErrorFlags::none(),
             )))
         ),
         case::address_v4_format(
@@ -972,7 +976,7 @@ mod tests {
                     value: Value::Address(IpAddr::V4(Ipv4Addr::new(10, 10, 0, 1))),
                     padding: Vec::new(),
                 },
-                ErrorFlags::NON_ZERO_RESERVED,
+                ErrorFlags::NonZeroReserved.into(),
             )))
         ),
         case::address_v6_format(
@@ -1005,7 +1009,7 @@ mod tests {
                                 0x0000, 0x8a2e, 0x0370, 0x7334))),
                     padding: Vec::new(),
                 },
-                ErrorFlags::NONE,
+                ErrorFlags::none(),
             )))
         ),
         case::time_format(
@@ -1033,7 +1037,7 @@ mod tests {
                     value: Value::Time(3_794_601_600),
                     padding: Vec::new(),
                 },
-                ErrorFlags::NONE,
+                ErrorFlags::none(),
             )))
         ),
         case::grouped_format(
@@ -1111,7 +1115,7 @@ mod tests {
                         }]),
                     padding: Vec::new(),
                 },
-                ErrorFlags::NON_ZERO_PADDING | ErrorFlags::NON_ZERO_RESERVED
+                ErrorFlags::NonZeroPadding | ErrorFlags::NonZeroReserved
             )))
         ),
         case::invalid_utf8(
@@ -1139,7 +1143,7 @@ mod tests {
                     value: Value::Unhandled(vec![0xfe, 0xfe, 0xff, 0xff]),
                     padding: Vec::new(),
                 },
-                ErrorFlags::DATA_VALUE,
+                ErrorFlags::DataValue.into(),
             )))
         ),
         case::invalid_address(
@@ -1169,7 +1173,7 @@ mod tests {
                     value: Value::Unhandled(vec![0x0a, 0x0a, 0x00, 0x01, 0x01]),
                     padding: vec![0x00, 0x00, 0x00],
                 },
-                ErrorFlags::DATA_LENGTH,
+                ErrorFlags::DataLength.into(),
             )))
         ),
         case::unhandled(
@@ -1199,11 +1203,11 @@ mod tests {
                     value: Value::Unhandled(vec![0x0a, 0x0a, 0x00, 0x01, 0x01]),
                     padding: vec![0x00, 0x00, 0x00],
                 },
-                ErrorFlags::NONE,
+                ErrorFlags::none(),
             )))
         )
     )]
-    fn test_avp(input: &[u8], expected: IResult<&[u8], (AVP, ErrorFlags)>) {
+    fn test_avp(input: &[u8], expected: IResult<&[u8], (AVP, Flags<ErrorFlags>)>) {
         assert_eq!(AVP::parse(input), expected);
     }
 
@@ -1240,7 +1244,7 @@ mod tests {
                         end_id: 0x7dc0_a11b,
                     },
                     avps: Vec::new(),
-                    error_flags: ErrorFlags::NONE,
+                    error_flags: ErrorFlags::none(),
                 })
             ))
         ),
@@ -1323,7 +1327,7 @@ mod tests {
                         padding: Vec::new(),
                     },
                 ],
-                error_flags : ErrorFlags::NON_ZERO_RESERVED | ErrorFlags::NON_ZERO_PADDING,
+                error_flags : ErrorFlags::NonZeroReserved | ErrorFlags::NonZeroPadding,
             })
         ))),
         case::incomplete(